Hide meaningless options in Wi-Fi setup
Based on if configured as AP or client, some entries in page 2 make
sense to be configured.
I.e. the use of a DHCP makes only sense when it is a client and
selecting the channel only makes sense when it is an AP.
Further, when in client mode and with enabled DHCP, configuration of IP
etc. doesn't make sense.