add declarative firmware build via nix
This allows developers with a linux machine or docker or wsl to build
the app reproducibly via this command.

```nix
nix build .#goggle-app
```

Except the nix package manager, no other tools need to be installed and
no environment needs to be set up.

The nix build system is quite flexible and has a rich ecosystem.
In the future, the nix declarations could be extended in order to:

- add automated testing to the build pipeline
- switch to a recent toolchain instead of the outdated one from 2018
currently pulled in by `setup.sh`
- cleanup the repo from binary files like busybox, tinycurl, dropbear,
etc. Instead declare dependencies through its source and build
definition, which simplifies updating, modifying and adding stuff.
- update the currently used 4.9.X linux kernel which is end-of-life
since 2023 and likely insecure
- declare a reproducible build for the base os, so it can be more easily
extended by the community
- improve the firmware architecture and update mechanism to make it easy
to deliver updates including the base os